Made pickit a module (big commit). This will break stuff, i think. Development commit
